Add `@Symbol` to CmakeTool's descriptor
That should make it possible to use an auto-installed CMake using a
`tool` section in a pipeline, so a normal `sh`/`bat`/`pwsh` step can use
it freely.

This also makes the descriptor implement PersistentDescriptor and
overrides setInstallations(). I'm not sure whether this is actually
required, but it's probably harmless either way.